[FLINK-21092][table-planner-blink] Move the spec classes in `exec.utils` to `exec.spec` package
