+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/spec/IntervalJoinSpec.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.plan.nodes.exec.utils;
+package org.apache.flink.table.planner.plan.nodes.exec.spec;
 
 /**
  * IntervalJoinSpec describes how two tables will be joined in interval join.
