+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/spec/JoinSpec.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.plan.nodes.exec.utils;
+package org.apache.flink.table.planner.plan.nodes.exec.spec;
 
 import org.apache.flink.table.runtime.operators.join.FlinkJoinType;
 import org.apache.flink.util.Preconditions;
